requirements.txt
# 华为云优化分析项目依赖

# 核心依赖
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0

# 环境管理
python-dotenv>=1.0.0

# 日期处理
python-dateutil>=2.8.0

# 数据验证
numpy>=1.24.0